% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions.R
\name{actions_pause}
\alias{actions_pause}
\title{Wait for a period of time}
\usage{
actions_pause(seconds)
}
\arguments{
\item{seconds}{The number of seconds to wait for.}
}
\value{
A \code{selenium_action} object.
}
\description{
A pause action to be passed into \code{\link[=actions_stream]{actions_stream()}}. Waits for a given
number of seconds before performing the next action in the stream.
}
\examples{
actions_stream(
  actions_pause(1)
)

}
