% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions.R
\name{actions_press}
\alias{actions_press}
\alias{actions_release}
\title{Press or release a key}
\usage{
actions_press(key)

actions_release(key)
}
\arguments{
\item{key}{The key to press: a string consisting of a single character. Use
the \link{keys} object to use special keys (e.g. \code{Ctrl}).}
}
\value{
A \code{selenium_action} object.
}
\description{
Key actions to be passed into \code{\link[=actions_stream]{actions_stream()}}. \code{actions_press()}
represents pressing a key on the keyboard, while \code{actions_release()}
represents releasing a key.
}
\examples{
actions_stream(
  actions_press("a"),
  actions_release("a"),
  actions_press(keys$enter),
  actions_release(keys$enter)
)

}
