% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keys.R
\docType{data}
\name{keys}
\alias{keys}
\title{A list of special keys}
\format{
An object of class \code{list} of length 65.
}
\usage{
keys
}
\description{
A named list of special keys, where each key is a single Unicode character,
which will be interpreted by selenium as a special key. Each key is just a
string, so can be used with string manipulaton functions like \code{\link[=paste]{paste()}}
without any special treatment.
}
\keyword{datasets}
